requirements.txt
Django>=1.8,<1.9
psycopg2
djangorestframework
django_compressor
